add some changes to docker test cases
base_test.py and test_docker_client.py
