+++ tests/base_tests.py
@@ -1,6 +1,7 @@
 from unittest import TestCase
 from loguru import logger
 from uuid import uuid4
+from subprocess import Popen, PIPE, STDOUT
 
 
 class BaseTests(TestCase):
@@ -19,3 +20,9 @@ class BaseTests(TestCase):
 
     def info(self, message):
         BaseTests.LOGGER.info(message)
+
+    @staticmethod
+    def os_command(command):
+        process = Popen(command, shell=True, stdout=PIPE, stderr=STDOUT, universal_newlines=True)
+        output, error = process.communicate()
+        return output, error
